- moved ToViewModel(this Location location) and ToModel(this LocationViewModel location) out of EventViewModelExtensions class and into newly created LocationExtensions and LocationViewModelsExtensions classes (note: from what ReSharper is tell me, neither of these methods are actually used, so might be worth deleting them later) -renamed EventViewModelExtension class to EventViewModelExtensionS
