+++ AllReadyApp/Web-App/AllReady/ViewModels/Event/EventViewModel.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Security.Claims;
 using AllReady.Models;
 using AllReady.ViewModels.Shared;
 using AllReady.ViewModels.Task;
@@ -94,67 +93,12 @@ namespace AllReady.ViewModels.Event
         public bool HasHeadline => !string.IsNullOrEmpty(Headline);
     }
 
-    public static class EventViewModelExtension
+    public static class EventViewModelExtensions
     {
-        public static LocationViewModel ToViewModel(this Location location)
-        {
-            var value = new LocationViewModel
-            {
-                Address1 = location.Address1,
-                Address2 = location.Address2,
-                City = location.City,
-                PostalCode = location.PostalCode,
-                State = location.State
-            };
-            return value;
-        }
-
-        public static Location ToModel(this LocationViewModel location)
-        {
-            var value = new Location
-            {
-                Address1 = location.Address1,
-                Address2 = location.Address2,
-                City = location.City,
-                PostalCode = location.PostalCode,
-                State = location.State,
-                Country = "TODO:  Put country in both objects"
-            };
-            return value;
-        }
-
+        //this method is really operating on the Event class as it's extension type, so must like should be moved to a new EventExtensions class and kept with the Event.cs file
         public static IEnumerable<EventViewModel> ToViewModel(this IEnumerable<Models.Event> campaignEvents)
         {
             return campaignEvents.Select(campaignEvent => new EventViewModel(campaignEvent));
         }
-
-        //public static EventViewModel WithUserInfo(this EventViewModel viewModel, Models.Event campaignEvent, ClaimsPrincipal user, IAllReadyDataAccess dataAccess)
-        //{
-        //    if (user.IsSignedIn())
-        //    {
-        //        var userId = user.GetUserId();
-        //        var appUser = dataAccess.GetUser(userId);
-        //        viewModel.UserId = userId;
-        //        viewModel.UserSkills = appUser?.AssociatedSkills?.Select(us => new SkillViewModel(us.Skill)).ToList();
-        //        viewModel.IsUserVolunteeredForEvent = dataAccess.GetEventSignups(viewModel.Id, userId).Any();
-        //        var assignedTasks = campaignEvent.Tasks.Where(t => t.AssignedVolunteers.Any(au => au.User.Id == userId)).ToList();
-        //        viewModel.UserTasks = new List<TaskViewModel>(assignedTasks.Select(data => new TaskViewModel(data, userId)).OrderBy(task => task.StartDateTime));
-        //        var unassignedTasks = campaignEvent.Tasks.Where(t => t.AssignedVolunteers.All(au => au.User.Id != userId)).ToList();
-        //        viewModel.Tasks = new List<TaskViewModel>(unassignedTasks.Select(data => new TaskViewModel(data, userId)).OrderBy(task => task.StartDateTime));
-        //        viewModel.SignupModel = new EventSignupViewModel()
-        //        {
-        //            EventId = viewModel.Id,
-        //            UserId = userId,
-        //            Name = appUser.Name,
-        //            PreferredEmail = appUser.Email,
-        //            PreferredPhoneNumber = appUser.PhoneNumber
-        //        };
-        //    }
-        //    else
-        //    {
-        //        viewModel.UserTasks = new List<TaskViewModel>();
-        //    }
-        //    return viewModel;
-        //}
     }
 }
